change percent_off to float. tests pass

+++ db/migrate/20200229154922_create_discount.rb
@@ -1,7 +1,7 @@
 class CreateDiscount < ActiveRecord::Migration[5.1]
   def change
     create_table :discounts do |t|
-      t.integer :percent_off
+      t.float :percent_off
       t.integer :quantity_threshold
       t.integer :status, default: 0
       t.datetime "created_at", null: false
